requirements.txt
blinker
certifi
cffi
charset-normalizer
click
cryptography
dominate
Flask
Flask-Bootstrap
Flask-SQLAlchemy
Flask-WTF
idna
itsdangerous
Jinja2
PyJWT
MarkupSafe
pycparser
requests
SQLAlchemy
typing_extensions
urllib3
visitor
Werkzeug
WTForms
